requirements.txt
Django==1.11
django-model-utils==3.1.2
pytz==2018.7
